[feature/#62]Exception 처리 및 파라미터 Positive처리
