+++ src/main/java/com/example/jdnc_library/feature/book/service/BookService.java
@@ -7,19 +7,14 @@ import com.example.jdnc_library.domain.book.model.CollectionInfo;
 import com.example.jdnc_library.domain.book.repository.BookRepository;
 import com.example.jdnc_library.domain.book.repository.BorrowRepository;
 import com.example.jdnc_library.domain.book.repository.CollectionRepository;
-import com.example.jdnc_library.domain.member.model.Member;
 import com.example.jdnc_library.exception.clienterror._400.EntityNotFoundException;
-import com.example.jdnc_library.exception.clienterror._400.ExistEntityException;
-import com.example.jdnc_library.feature.book.DTO.*;
-import com.example.jdnc_library.feature.book.repository.BookInfoQueryRepository;
+import com.example.jdnc_library.feature.book.DTO.BookRequest;
+import com.example.jdnc_library.feature.book.DTO.BorrowListDTO;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -30,159 +25,7 @@ public class BookService {
     private final BookRepository bookRepository;
     private final CollectionRepository collectionRepository;
     private final BorrowRepository borrowRepository;
-    private final BookInfoQueryRepository bookInfoQueryRepository;
 
-    /**
-     * 책 1권의 정보를 리턴
-     * @param id
-     * @return
-     */
-    public BookDetailDTO getBookById(long id){
-        Optional<BookInfo> book = bookRepository.findById(id);
-        if (book.isPresent()) {
-            BookInfo newBookInfo = book.get();
-            return new BookDetailDTO(id, newBookInfo.getTitle(), newBookInfo.getImage(), newBookInfo.getContent(), newBookInfo.getAuthor(), newBookInfo.getPublisher(), newBookInfo.getBookGroup(), available(newBookInfo.getId()));
-        }
-        else {
-            return null;
-        }
-    }
-
-    /**
-     * 대출 가능 여부 확인
-     * @param id
-     * @return
-     */
-    public boolean available(long id){
-        List<CollectionInfo> collectionInfo = collectionRepository.findByBookInfo_id(id);
-
-        return collectionInfo.stream().anyMatch(CollectionInfo::isAvailable);
-    }
-
-    /**
-     * 제목으로 검색했을 때 검색어가 포함된 모든 항목을 리턴
-     * @param title
-     * @return
-     */
-    public List<BookListDTO> searchBooks(String title, Pageable pageable) {
-        Page<CollectionInfo> collectionPage = collectionRepository.findAllByBookInfo_TitleContaining(title, pageable);
-
-        return collectionPage.getContent()
-                .stream()
-                .map(BookListDTO::of)
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * @param title
-     * @param pageable
-     * @return
-     */
-    //TODO : 이부분 컨트롤러 까지 수정 - by KBJ
-    public List<BookListDTO> getBookListDTOs(String title, Pageable pageable){
-        return bookInfoQueryRepository.getBookListDTOs(title, pageable).getContent();
-    }
-
-    /**
-     * 모든 책 리턴
-     * @return
-     */
-    public List<BookListDTO> searchAllBooks(Pageable pageable){
-        Page<BookInfo> bookInfoList = bookRepository.findAll(pageable);
-
-        return bookInfoList.getContent()
-                .stream()
-                .map(bookInfo -> {
-                    boolean availableForBorrow = available(bookInfo.getId());
-                    return BookListDTO.of(bookInfo, availableForBorrow);
-                })
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * 그룹으로 나눠서 검색
-     * @param group
-     * @return
-     */
-    public List<BookListDTO> searchBooksByGroup(String group, Pageable pageable){
-        Page<BookInfo> bookInfoList = bookRepository.findByBookGroup(group, pageable);
-
-        return bookInfoList.getContent()
-                .stream()
-                .map(bookInfo -> {
-                    boolean availableForBorrow = available(bookInfo.getId());
-                    return BookListDTO.of(bookInfo, availableForBorrow);
-                })
-                .collect(Collectors.toList());
-    }
-
-    public List<BookListDTO> searchBooksByGroupAndTitle(String group, String title,Pageable pageable){
-        Page<BookInfo> bookInfoList = bookRepository.findByBookGroupAndTitleContaining(group, title, pageable);
-
-        return bookInfoList.getContent()
-                .stream()
-                .map(bookInfo -> {
-                    boolean availableForBorrow = available(bookInfo.getId());
-                    return BookListDTO.of(bookInfo, availableForBorrow);
-                })
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * 책의 QR을 찍었을 때 책 정보 리턴
-     * @param bookNumber
-     * @return
-     */
-    public BorrowDetailDTO qrBook(long bookNumber){
-        Optional<CollectionInfo> collectionInfoOptional = collectionRepository.findByBookNumber(bookNumber);
-        CollectionInfo collectionInfo = collectionInfoOptional.get();
-        return BorrowDetailDTO.of(collectionInfo);
-    }
-
-    /**
-     * 책을 빌릴 때 데이터베이스에 저장
-     * @param bookNumber
-     */
-    //TODO : 중복 대여 처리하기
-    @Transactional
-    public void borrowBook(long bookNumber){
-        Optional<BorrowInfo> borrowInfoOptional = borrowRepository.findByCollectionInfo_BookNumberAndReturnDateIsNull(bookNumber);
-        if(borrowInfoOptional.isPresent()){
-            throw new ExistEntityException();
-        }
-        else {
-            CollectionInfo collectionInfo = collectionRepository.findByBookNumber(bookNumber)
-                    .orElseThrow(() -> new EntityNotFoundException(bookNumber, CollectionInfo.class));
-
-            BorrowInfo borrowInfo = new BorrowInfo(collectionInfo);
-            borrowRepository.save(borrowInfo);
-            collectionInfo.updateAvailable(false);
-        }
-    }
-
-    /**
-     * 반납함 QR을 찍었을 때 반납 가능한 도서 리스트를 리턴
-     * @param member
-     * @return
-     */
-    public List<BorrowListDTO> returnBookList(Member member, Pageable pageable){
-        List<BorrowInfo> borrowInfoList = borrowRepository.findAllByCreatedByAndReturnDateIsNull(member, pageable).getContent();
-
-        return borrowInfoList.stream()
-                .map(BorrowListDTO::of)
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * 책 한권을 반납하면 데이터베이스에 저장
-     * @param bookNumber
-     */
-    public void returnBook(long bookNumber){
-        Optional<BorrowInfo> borrowInfoOptional = borrowRepository.findByCollectionInfo_BookNumber(bookNumber);
-        BorrowInfo borrowInfo = borrowInfoOptional.get();
-        borrowInfo.returnBook(LocalDateTime.now());
-        borrowRepository.save(borrowInfo);
-    }
 
     /**
      * 입력한 달에 빌린 책 리스트를 리턴(Admin)
